+++ src/main/java/christmas/domain/discount/Discount.java
@@ -1,6 +1,5 @@
 package christmas.domain.discount;
 
-import christmas.domain.VisitDate;
 import christmas.domain.menu.component.Price;
 
 public interface Discount {
