+++ src/main/java/christmas/domain/discount/SpecialDiscount.java
@@ -1,8 +1,9 @@
 package christmas.domain.discount;
 
-import christmas.domain.VisitDate;
+import christmas.domain.date.VisitDate;
+import christmas.domain.date.DecemberCalendar;
 import christmas.domain.menu.component.Price;
-import christmas.exception.discount.InvalidStarException;
+import christmas.exception.inside.discount.InvalidStarException;
 
 public class SpecialDiscount implements Discount{
