--- src/main/java/christmas/exception/NotExistsMenuException.java
@@ -1,8 +1,0 @@
-package christmas.exception;
-
-public class NotExistsMenuException extends IllegalArgumentException{
-    private static final String MESSAGE = "[ERROR] 유효하지 않은 주문입니다. 다시 입력해 주세요.";
-    public NotExistsMenuException() {
-        super(MESSAGE);
-    }
-}
