+++ src/main/java/christmas/exception/inside/discount/InvalidWeekendException.java
@@ -1,4 +1,4 @@
-package christmas.exception.discount;
+package christmas.exception.inside.discount;
 
 public class InvalidWeekendException extends IllegalArgumentException{
     private static final String MESSAGE = "[ERROR] 주말이 아닌데, 주말 할인을 호출했습니다. (내부 오류)";
