+++ src/main/java/christmas/exception/outside/DateUnitException.java
@@ -1,4 +1,4 @@
-package christmas.exception;
+package christmas.exception.outside;
 
 public class DateUnitException extends IllegalArgumentException {
     private static final String MESSAGE = "[ERROR] 유효하지 않은 날짜입니다. 다시 입력해 주세요.";
