+++ src/main/java/christmas/exception/outside/NotExistsMenuException.java
@@ -1,8 +1,8 @@
-package christmas.exception;
+package christmas.exception.outside;
 
-public class PriceUnitException extends IllegalArgumentException {
+public class NotExistsMenuException extends IllegalArgumentException{
     private static final String MESSAGE = "[ERROR] 유효하지 않은 주문입니다. 다시 입력해 주세요.";
-    public PriceUnitException() {
+    public NotExistsMenuException() {
         super(MESSAGE);
     }
 }
