+++ src/main/java/christmas/exception/outside/order/OrderOnlyDrinkException.java
@@ -1,4 +1,4 @@
-package christmas.exception.order;
+package christmas.exception.outside.order;
 
 public class OrderOnlyDrinkException extends IllegalArgumentException {
     private static final String MESSAGE = "[ERROR] 유효하지 않은 주문입니다. 다시 입력해 주세요.";
