+++ src/test/java/christmas/domain/discount/ChristmasDiscountTest.java
@@ -3,9 +3,9 @@ package christmas.domain.discount;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import christmas.domain.VisitDate;
+import christmas.domain.date.VisitDate;
 import christmas.domain.menu.component.Price;
-import christmas.exception.discount.XmasDiscountUnitException;
+import christmas.exception.inside.discount.XmasDiscountUnitException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
