+++ src/test/java/christmas/domain/order/OrdersTest.java
@@ -3,13 +3,13 @@ package christmas.domain.order;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import christmas.domain.VisitDate;
-import christmas.exception.DateUnitException;
-import christmas.exception.NotExistsMenuException;
-import christmas.exception.order.OrderDuplicateMenuException;
-import christmas.exception.order.OrderOnlyDrinkException;
-import christmas.exception.order.OrderTotalExceededException;
-import christmas.exception.order.OrderValueMinUnitException;
+import christmas.domain.date.VisitDate;
+import christmas.exception.outside.DateUnitException;
+import christmas.exception.outside.NotExistsMenuException;
+import christmas.exception.outside.order.OrderDuplicateMenuException;
+import christmas.exception.outside.order.OrderOnlyDrinkException;
+import christmas.exception.outside.order.OrderTotalExceededException;
+import christmas.exception.outside.order.OrderValueMinUnitException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
